Renamed flyercode to flyer.js; Added fields from old caseDetails page
