--- client/src/components/case-upload/flyercode.js
@@ -1,109 +1,0 @@
-import React, {Component} from 'react';
-import QRCode from 'qrcode.react';
-import { FacebookShareButton, FacebookIcon, GooglePlusShareButton, GooglePlusIcon, TwitterShareButton, TwitterIcon } from 'react-share';
-import './upload.scss';
-
-import placeholderImage from '../../assets/images/cover1.jpg';
-
-class FlyerCode extends Component {
-
-    printFlyer = () => {
-        window.print();
-        return false;
-    }
-
-    render() {
-        const { caseid } = this.props.match.params;
-
-        if (!caseid) {
-            return <div>An error occurred. CaseId is required in the URL.</div>
-        }
-
-        const url = `http://pawsfindhome.com/casedetails/${caseid}`;
-
-        const data = {
-            animalType: 'dog',
-            caseType: 'found',
-            caseDate: '3/13/2019',
-            phone: '1-714-604-5096',
-            location: {
-                location: 'Irvine Spectrum Dr & Mullen',
-                city: 'Irvine',
-                state: 'CA'
-            },
-            description: 'Tan seeing-eye dog.  Answers to Ralph.  My best friend.'
-        };
-
-        const {location, city, state} = data.location;
-        const animalType = data.animalType === 'other' ? 'pet' : data.animalType;
-        const caseType = data.caseType[0].toUpperCase() + data.caseType.slice(1);
-        const socialMediaTitle = data.caseType === 'found' ?
-            `Please help me identify found ${animalType}` :
-            `Please help find my ${animalType}`;
-
-        const description = `${caseType} on ${data.caseDate}. ${data.description}`
-        const locationText = `${location}, ${city} ${state}`;
-
-        return (
-            <div id="flyer">
-                <div className="buttons center">
-                    <FacebookShareButton url={url}>
-                        <FacebookIcon size={40} round />
-                    </FacebookShareButton>
-                    <GooglePlusShareButton url={url}>
-                        <GooglePlusIcon size={40} round />
-                    </GooglePlusShareButton>
-                    <TwitterShareButton url={url} title={socialMediaTitle}>
-                        <TwitterIcon size={40} round />
-                    </TwitterShareButton>
-                    <button onClick={this.printFlyer} className="btn-floating"><i className="material-icons">print</i></button>
-                </div>
-                <div className="row">
-                    <div className="col s10 offset-s1 center title">
-                        <h2>{caseType.toUpperCase()} {animalType.toUpperCase()}</h2>
-                    </div>
-                </div>
-                <div className="row">
-                    <div className="col s8 offset-s2 m4 offset-m4 center image-container">
-                        <img src={placeholderImage} className="responsive-img"/>
-                    </div>
-                </div>
-                <div className="row">
-                    <div className="col s12 bottom">
-                        <div className="details">
-                            <table>
-                                <tbody>
-                                    <tr><td><i className="material-icons">phone</i> {data.phone}</td></tr>
-                                    <tr><td><i className="material-icons">location_on</i> {locationText}</td></tr>
-                                </tbody>
-                            </table>
-                        </div>
-                        <div className="qrcode">
-                            <QRCode
-                                value={url}
-                                size={128}
-                                fgColor='#000000'
-                                bgColor='#ffffff'
-                                level='L'
-                                renderAs='svg'
-                                includeMargin={true}
-                                />
-                        </div>
-                    </div>
-                </div>
-                <div className="row">
-                    <div className="col s12">
-                    DESCRIPTION: <span className="description">{description}</span>
-                    </div>
-                </div>
-                <div className="row">
-                    <div className="col s12 center">
-                    Generated by PAWSFindHome.com
-                    </div>
-                </div>
-          </div>
-        )
-    }
-}
-
-export default FlyerCode;
